+++ src/app/shell/personal-cabinet/parent/parent.module.ts
@@ -19,8 +19,6 @@ import { SharedModule } from 'src/app/shared/shared.module';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { FavoriteWorkshopsComponent } from './favorite-workshops/favorite-workshops.component';
 
-
-
 @NgModule({
   declarations: [
     CreateChildComponent,
