+++ src/app/shell/workshop-details/workshop-details.module.ts
@@ -3,8 +3,7 @@ import { WorkshopPageComponent } from './workshop-page/workshop-page.component';
 import { SideMenuComponent } from './side-menu/side-menu.component';
 import { WorkshopDetailsComponent } from './workshop-details.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { SharedModule } from 'src/app/shared/shared.module';
-import { MaterialModule } from 'src/app/shared/material/material.module';
+import { MaterialModule } from 'src/app/shared/modules/material.module';
 import { RouterModule } from '@angular/router';
 import { WorkshopAboutComponent } from './workshop-page/workshop-about/workshop-about.component';
 import { ProviderAboutComponent } from './workshop-page/provider-about/provider-about.component';
@@ -15,6 +14,7 @@ import { ActionsComponent } from './side-menu/actions/actions.component';
 import { ContactsComponent } from './side-menu/contacts/contacts.component';
 import { ScheduleComponent } from './side-menu/schedule/schedule.component';
 import { AllProviderWorkshopsComponent } from './workshop-page/all-provider-workshops/all-provider-workshops.component';
+import { SharedModule } from 'src/app/shared/shared.module';
 
 @NgModule({
   declarations: [
